requirements.txt
PyYAML==5.3.1
Flask==1.1.2
python-dotenv==0.13.0
